fuzzer: made fuzzer tools more generic that it works more systems
The main entrypoint is now the ./fuzz.sh script located in test/fuzzer/.
It can be used to both setup the required data to run the fuzzer and to
run the fuzzer itself. The script is able to run without any additional
flags if the working directory is test/fuzzer. Otherwise at least the
--project-root flag will have to be set. (Invoke -h/--help to see the
other flags).