fuzzer: Added *a lot* of strings to the dictionary corups
Each section of strings is labled with the file from where they
originated. I might have accidentaly removed some relevant ones and
there are likely a lot of irrelevant ones still in there, but I don't
really think that matters that much. There are also probably a few
duplicates in there. It als might be usful tho to prefix some commands
with \\, but lets see :).