Fixed `insertAtomIntoCol` crash described in #163
I'm not quite sure if this is fully valid, but if I understood
correctly the issue with this is when calling std::vector<T>::insert
with an iterator outside the size of vector causes either a simple
overflow or more likely it'll resize the vector without initializing the
values before the inserted value. This patch addresses this by expanding
the vector to the needed size beforehand.

Reproducible using:
- \begin{cases}&{}\\\\